+++ frontend/src/pages/oj/views/index.js
@@ -31,4 +31,3 @@ export {
 export * from './classroom'
 export * from './languagepack'
 export * from './review'
-export * from './career'